+++ lib/main.dart
@@ -1,18 +1,14 @@
+import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
+import 'package:scdao_mobile/screens/main_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:scdao_mobile/models/user.dart';
 import 'package:scdao_mobile/providers/user.dart';
-import 'package:scdao_mobile/screens/new_document_screen.dart';
 import 'package:scdao_mobile/screens/signup_screen.dart';
 import 'package:scdao_mobile/screens/login_screen.dart';
-import 'package:scdao_mobile/screens/document_screen.dart';
 import 'package:scdao_mobile/screens/review_screen.dart';
-import 'package:scdao_mobile/screens/Settings_screen.dart';
-import 'package:scdao_mobile/screens/Privacy/privacy_screen.dart';
-import 'package:scdao_mobile/screens/Privacy/change_password.dart';
-import 'package:scdao_mobile/screens/Privacy/change_username.dart';
 import 'package:scdao_mobile/services/user.dart';
 import 'dart:async';
 
@@ -20,6 +16,11 @@ Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   SystemChrome.setEnabledSystemUIMode(SystemUiMode.manual,
       overlays: [SystemUiOverlay.top]);
+
+  // get camera
+  // final camera = await availableCameras();
+  // final firstCam = camera.first;
+
   runApp(MultiProvider(
     providers: [
       ChangeNotifierProvider(create: (context) => UserProvider()),
@@ -29,9 +30,15 @@ Future<void> main() async {
       builder: ((context, AsyncSnapshot<SharedPreferences> snapshot) {
         switch (snapshot.connectionState) {
           case ConnectionState.done:
-            return MyApp(prefs: snapshot.data);
+            if (snapshot.data != null)
+              return MyApp(
+                prefs: snapshot.data!,
+                // camera: firstCam,
+              );
+            return MaterialApp(
+                home: LoadingScreen()); //TODO: create error screen
           default:
-            return LoadingScreen();
+            return MaterialApp(home: LoadingScreen());
         }
       }),
     ),
@@ -39,8 +46,9 @@ Future<void> main() async {
 }
 
 class MyApp extends StatefulWidget {
-  final SharedPreferences? prefs;
-  MyApp({Key? key, this.prefs}) : super(key: key);
+  late final SharedPreferences prefs;
+  // final CameraDescription camera;
+  MyApp({Key? key, required this.prefs}) : super(key: key);
 
   @override
   State<MyApp> createState() => _MyAppState();
@@ -52,15 +60,14 @@ class _MyAppState extends State<MyApp> {
   // get token from cache and sends it to backend to check
   // if it's still valid or not
   Future<void> checkCachedToken(UserProvider provider) async {
-    if (widget.prefs == null) return;
-    final token = TokenModel.getTokenFromCache(widget.prefs!);
+    final token = TokenModel.getTokenFromCache(widget.prefs);
     if (token != null) {
       final user = await userService.checkToken(token);
       if (user != null) {
         provider.setUser(user);
       }
     } else {
-      await widget.prefs!.clear();
+      await widget.prefs.clear();
     }
   }
 
@@ -75,18 +82,15 @@ class _MyAppState extends State<MyApp> {
 
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: NewDocumentScreen(),
+      initialRoute: "/",
       routes: {
+        "/": (_) => LoadingScreen(),
         LoginScreen.routeName: (_) => LoginScreen(
               prefs: widget.prefs,
             ),
         SignupScreen.routeName: (_) => SignupScreen(),
-        DocumentScreen.routeName: (_) => DocumentScreen(),
         ReviewScreen.routeName: (_) => ReviewScreen(),
-        SettingsScreen.routeName: (_) => SettingsScreen(),
-        PrivacyScreen.routeName: (_) => PrivacyScreen(),
-        UsernameScreen.routeName: (_) => UsernameScreen(),
-        PasswordScreen.routeName: (_) => PasswordScreen(),
+        MainScreen.routeName: (_) => MainScreen(),
       },
     );
   }
@@ -98,8 +102,30 @@ class LoadingScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Text("Loading"),
+    return Scaffold(
+      appBar: AppBar(
+        title: Text("Loading screen"),
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            const Text(
+              "Loading",
+              style: TextStyle(
+                fontSize: 50.0,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            IconButton(
+              icon: Icon(
+                Icons.document_scanner,
+              ),
+              onPressed: () =>
+                  Navigator.of(context).pushReplacementNamed("/documents"),
+            ),
+          ],
+        ),
+      ),
     );
   }
 }
