+++ lib/screens/Settings_screen.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/gestures.dart';
 
 class SettingsScreen extends StatefulWidget {
-  static const String routeName = "/settings";
+  static const String routeName = "documents/settings";
   @override
   _SettingsScreenState createState() => _SettingsScreenState();
 }
