+++ lib/screens/review_screen.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 class ReviewScreen extends StatefulWidget {
