+++ pubspec.yaml
@@ -27,9 +27,7 @@ dependencies:
     sdk: flutter
   path_provider:
   path:
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  camera:
   cupertino_icons: ^1.0.2
   flutter_platform_widgets: ^2.0.0
   font_awesome_flutter: ^10.1.0
